release 0.1.183 source code

+++ services/workspace/v2/model/model_volume.go
@@ -9,11 +9,11 @@ import (
 // Volume 磁盘。
 type Volume struct {
 
-	// 桌面数据盘对应的磁盘类型，需要与系统所提供的磁盘类型相匹配。  -SAS：高IO。 -SSD：超高IO。 -GPSSD：通用型SSD。
+	// 磁盘类型。
 	Type string `json:"type"`
 
-	// 磁盘容量，单位GB。系统盘大小范围[80-32760]，数据盘范围[10-32760]，大小为10的倍数。
-	Size int32 `json:"size"`
+	// 磁盘大小，单位GB。
+	Size int64 `json:"size"`
 }
 
 func (o Volume) String() string {
